Add buildSrc module and java conventions.
