+++ examples/bukkit/build.gradle.kts
@@ -1,11 +1,13 @@
 plugins {
     id("java")
-    id("com.github.johnrengelman.shadow")
+    id("com.github.johnrengelman.shadow") version "7.1.2"
     id("net.minecrell.plugin-yml.bukkit") version "0.5.2"
 }
 
 repositories {
     mavenCentral()
+    maven { url = uri("https://repo.panda-lang.org/releases/") }
+    maven { url = uri("https://hub.spigotmc.org/nexus/content/repositories/snapshots/") }
 }
 
 dependencies {
