+++ litecommands-bungee/build.gradle.kts
@@ -1,10 +1,14 @@
+plugins {
+    id("litecommands.java-conventions")
+}
+
 dependencies {
     api(project(":litecommands-core"))
 
     compileOnly("net.md-5:bungeecord-api:1.19-R0.1-SNAPSHOT")
 }
 
-val bungeeArtifact: String by rootProject.extra
+val bungeeArtifact: String by extra
 
 publishing {
     publications {
