+++ litecommands-core/src/main/java/dev/rollczi/litecommands/LiteCommandsExtension.java
@@ -1,0 +1,7 @@
+package dev.rollczi.litecommands;
+
+public interface LiteCommandsExtension {
+
+    void register(LiteCommandsBuilder<?> builder);
+
+}
