+++ litecommands-minestom/src/main/java/dev/rollczi/litecommands/minestom/LiteMinestomRegistryPlatform.java
@@ -1,16 +1,12 @@
 package dev.rollczi.litecommands.minestom;
 
-import dev.rollczi.litecommands.command.LiteInvocation;
 import dev.rollczi.litecommands.command.section.CommandSection;
-import dev.rollczi.litecommands.handle.ExecuteResultHandler;
 import dev.rollczi.litecommands.platform.ExecuteListener;
 import dev.rollczi.litecommands.platform.RegistryPlatform;
 import dev.rollczi.litecommands.platform.SuggestionListener;
-import net.minestom.server.MinecraftServer;
 import net.minestom.server.command.CommandManager;
 import net.minestom.server.command.CommandSender;
 import net.minestom.server.command.builder.Command;
-import net.minestom.server.network.socket.Server;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -20,12 +16,9 @@ class LiteMinestomRegistryPlatform implements RegistryPlatform<CommandSender> {
     private final Set<String> commands = new HashSet<>();
 
     private final CommandManager commandManager;
-    private ExecuteResultHandler<CommandSender> executeResultHandler = new ExecuteResultHandler<>();
-    private final boolean nativePermissions;
 
-    LiteMinestomRegistryPlatform(Server ignoredServer, boolean nativePermissions) {
-        this.nativePermissions = nativePermissions;
-        this.commandManager = MinecraftServer.getCommandManager();
+    LiteMinestomRegistryPlatform(CommandManager commandManager) {
+        this.commandManager = commandManager;
     }
 
     @Override
@@ -38,22 +31,7 @@ class LiteMinestomRegistryPlatform implements RegistryPlatform<CommandSender> {
     }
 
     private SimpleCommand createCommand(CommandSection<CommandSender> command, ExecuteListener<CommandSender> executeListener, SuggestionListener<CommandSender> suggestionListener) {
-        if (!this.nativePermissions) {
-            return new SimpleCommand(command, executeListener, suggestionListener);
-        }
-
-        MinestomNoPermission noPermissionHandler = (sender, requiredPermissions) -> {
-            LiteInvocation invocation = new LiteInvocation(new MinestomSender(sender), command.getName(), command.getName());
-
-            try {
-                this.executeResultHandler.handleResult(sender, invocation, requiredPermissions);
-                return true;
-            } catch (IllegalStateException exception) {
-                return false;
-            }
-        };
-
-        return new SimpleBlockedCommand(command, executeListener, suggestionListener, noPermissionHandler);
+        return new SimpleCommand(command, executeListener, suggestionListener);
     }
 
     @Override
@@ -72,14 +50,13 @@ class LiteMinestomRegistryPlatform implements RegistryPlatform<CommandSender> {
     public void unregisterAll() {
         for (String command : this.commands) {
             Command commandToDelete = this.commandManager.getCommand(command);
-            if(commandToDelete != null)
+
+            if (commandToDelete != null) {
                 this.commandManager.unregister(commandToDelete);
+            }
         }
 
         this.commands.clear();
     }
 
-    void setExecuteResultHandler(ExecuteResultHandler<CommandSender> executeResultHandler) {
-        this.executeResultHandler = executeResultHandler;
-    }
 }
