--- litecommands-minestom/src/main/java/dev/rollczi/litecommands/minestom/MinestomNoPermission.java
@@ -1,10 +1,0 @@
-package dev.rollczi.litecommands.minestom;
-
-import dev.rollczi.litecommands.command.permission.RequiredPermissions;
-import net.minestom.server.command.CommandSender;
-
-import java.util.function.BiFunction;
-
-interface MinestomNoPermission extends BiFunction<CommandSender, RequiredPermissions, Boolean> {
-
-}
