+++ settings.gradle.kts
@@ -5,10 +5,10 @@ include(":litecommands-velocity")
 include(":litecommands-bukkit")
 include(":litecommands-bukkit-adventure")
 include(":litecommands-bungee")
-includeModule(":litecommands-minestom", JavaVersion.VERSION_17)
+include(":litecommands-minestom", JavaVersion.VERSION_17)
 include(":examples:bukkit")
 
-fun includeModule(projectPath : String, version : JavaVersion) {
+fun include(projectPath : String, version : JavaVersion) {
     if (JavaVersion.current().isCompatibleWith(version)) {
         include(projectPath)
     }
